Adicionando activity de início e logo para equinos-qi
