+++ novaTentativa/app/src/main/res/layout/fragment_listar_testes.xml
@@ -17,16 +17,20 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/configuracaoRecycle"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="0dp"
+        android:layout_weight="9"
+        >
 
     </androidx.recyclerview.widget.RecyclerView>
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/floatingActionButton"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
+        android:layout_weight="1"
         android:layout_gravity="end|bottom"
         android:clickable="true"
+        android:foregroundGravity="center"
         android:src="@drawable/more_icon2"
         android:layout_marginRight="20sp"
          />
